+++ CHANGELOG.md
@@ -9,6 +9,13 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ---
 
+## [0.14.0] - (07-25-2022)
+
+- branch off NEXT connectivity from mgmtClient/f5Client
+  - Solves this issue:<https://github.com/f5devcentral/vscode-f5/issues/179>
+
+---
+
 ## [0.13.1] - (03-18-2022)
 
 - [bug] schema inject changes declaration to "dec" #20